+++ CHANGELOG.md
@@ -5,10 +5,13 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [0.1.0] - 2019-05-10
+- Removed the git_class module from this project and into its own module.
+
 ### Added
 - process_changes:  Process dirty and untracked files.
 
 ### Changed
+- Changed import of git_class to subdirectory of git_lib.
 - post_process:  Added Log class instance to function to record entries to log.
 - Added Log class instance for all post_process calls.
 - process_changes:  Replaced options for process_dirty and process_untracked to cfg settings.